backend/requirements.txt
Flask
APIFlask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
python-dotenv
marshmallow